move the load_source function
