+++ pandokia/common.py
@@ -18,23 +18,11 @@ from html import escape as html_escape
 from io import StringIO
 from urllib.parse import urlencode
 from urllib.parse import quote_plus
-import importlib.util
-import importlib.machinery
 
 import pandokia
 cfg = pandokia.cfg
 
-# replacement for imp.load_source() from https://docs.python.org/dev/whatsnew/3.12.html#imp
-# see similar usage in steuermann config.py
-def load_source(modname, filename):
-    loader = importlib.machinery.SourceFileLoader(modname, filename)
-    spec = importlib.util.spec_from_file_location(modname, filename, loader=loader)
-    module = importlib.util.module_from_spec(spec)
-    # The module is always executed and not cached in sys.modules.
-    # Uncomment the following line to cache the module.
-    # sys.modules[module.__name__] = module
-    loader.exec_module(module)
-    return module
+
 
 ######
 #--#--# CGI
