+++ pandokia/helpers/runner_unit2.py
@@ -1,7 +1,7 @@
 import sys
 import time
 import traceback
-from ..common import load_source
+from .importer import load_source
 import os.path
 
 no_unittest2 = False
